Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.pattern.cross-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
security.pattern.cross
Using an unclean pattern makes the target canvas origin-unclean, not the pattern canvas
Actual output:

PASS Using an unclean pattern makes the target canvas origin-unclean, not the pattern canvas
